requirements.txt
python-dotenv
pip~=23.0.1
setuptools~=67.6.1
wheel==0.37.1
python-irodsclient~=1.1.7
gen3~=4.9.4
Sphinx==7.0.1
sphinx-copybutton==0.5.2
sphinx-rtd-theme==1.2.2
sphinx-autodocgen==1.3
sphinx-autobuild==2021.3.14
PyYAML==5.4.1
openpyxl==3.1.2
pydicom~=2.4.3
h11~=0.9.0
click~=7.1.2
dictionaryutils~=3.0.0
fastavro~=1.7.3
attrs~=22.2.0
asyncio~=3.4.3
colorama~=0.4.6
numpy~=1.24.2
pandas~=1.5.3
nibabel~=5.1.0
requests~=2.28.2
cdislogging~=1.1.1
drsclient~=0.1.6
humanfriendly~=10.0
python-dateutil~=2.8.2
tqdm~=4.65.0
indexclient~=2.2.1
aiofiles~=0.8.0
aiohttp~=3.8.4
backoff~=1.11.1
jsonschema~=4.19.0
pytz~=2023.3
packaging~=23.0
idna~=3.4
multidict~=6.0.4
sniffio~=1.3.0
httpcore~=0.11.1
certifi~=2022.12.7
rfc3986~=1.5.0
scipy~=1.10.1
urllib3~=1.26.15
six~=1.16.0
xlrd~=2.0.1
psutil~=5.9.4
wcwidth~=0.2.6
defusedxml~=0.7.1
httpx~=0.15.5
zipp~=3.16.2
styleframe~=4.1
colour~=0.1.5
marshmallow~=3.19.0
pyreadline3~=3.4.1
gdcdictionary~=1.2.0
referencing~=0.30.2
pymongo~=3.11.2
dicom2nifti~=2.4.8

psycopg2
